[FEATURE] : CategoryVideo Retrofit을 가지고 오기 위한 HomeFragment 수정 및 CategoryVideAdapter 설정
[FIX] : 불필요한 gradle.app의 implementation 제거 (glide 중복) 및 추가(circleimageView)
