+++ app/build.gradle
@@ -40,19 +40,13 @@ android {
 dependencies {
     // spinner
     implementation "com.github.skydoves:powerspinner:1.2.6"
-
-/*    implementation 'com.google.apis:google-api-services-youtube:v3-rev183-1.22.0'
-    implementation 'com.google.http-client:google-http-client-android:1.38.1'
-    implementation 'com.google.api-client:google-api-client-android:1.31.2'
-    implementation 'com.google.api-client:google-api-client-gson:1.31.2'*/
-
     implementation ("com.google.code.gson:gson:2.10.1")
     implementation ("com.squareup.retrofit2:retrofit:2.9.0")
     implementation ("com.squareup.retrofit2:converter-gson:2.9.0")
     implementation ("com.squareup.okhttp3:okhttp:4.10.0")
     implementation ("com.squareup.okhttp3:logging-interceptor:4.10.0")
 
-    implementation ("com.github.bumptech.glide:glide:4.12.0")
+    implementation 'de.hdodenhof:circleimageview:3.1.0'
 
     implementation 'androidx.core:core-ktx:1.8.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
