+++ app/src/main/res/layout/fragment_my_video.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeFragment">
+    tools:context=".Home.Fragment.HomeFragment">
 
     <androidx.cardview.widget.CardView
         android:id="@+id/profile_card_view"
